net: move dns utility to net::
