+++ src/v/kafka/client/broker.cc
@@ -12,7 +12,7 @@
 #include "cluster/cluster_utils.h"
 #include "kafka/client/logger.h"
 #include "kafka/client/sasl_client.h"
-#include "rpc/dns.h"
+#include "net/dns.h"
 
 #include <seastar/core/coroutine.hh>
 #include <seastar/core/std-coroutine.hh>
