+++ src/v/net/CMakeLists.txt
@@ -2,6 +2,7 @@ v_cc_library(
   NAME net
   SRCS
     batched_output_stream.cc
+    dns.cc
   DEPS
     Seastar::seastar
   )
