+++ src/v/net/dns.h
@@ -8,14 +8,16 @@
  * the Business Source License, use of this software will be governed
  * by the Apache License, Version 2.0
  */
-
 #pragma once
 #include "utils/unresolved_address.h"
-namespace rpc {
+
+namespace net {
+
 /**
  * Resolves addresses using seastar DNS resolver. It uses mutex to workaround
  * seastar bug causing segmentation fault when udp channel is being accessed by
  * different fibers.
  */
 ss::future<ss::socket_address> resolve_dns(unresolved_address);
-} // namespace rpc
+
+} // namespace net
