+++ src/v/redpanda/admin_server.cc
@@ -37,7 +37,7 @@
 #include "redpanda/admin/api-doc/raft.json.h"
 #include "redpanda/admin/api-doc/security.json.h"
 #include "redpanda/admin/api-doc/status.json.h"
-#include "rpc/dns.h"
+#include "net/dns.h"
 #include "security/scram_algorithm.h"
 #include "security/scram_authenticator.h"
 #include "vlog.h"
@@ -257,7 +257,7 @@ ss::future<> admin_server::configure_listeners() {
                   });
             }
         }
-        auto resolved = co_await rpc::resolve_dns(ep.address);
+        auto resolved = co_await net::resolve_dns(ep.address);
         co_await ss::with_scheduling_group(_cfg.sg, [this, cred, resolved] {
             return _server.listen(resolved, cred);
         });
