+++ src/v/rpc/CMakeLists.txt
@@ -11,7 +11,6 @@ v_cc_library(
     reconnect_transport.cc
     connection_cache.cc
     simple_protocol.cc
-    dns.cc
   DEPS
     Seastar::seastar
     v::bytes
